Move save problem logic to menu
